vagon/requirements.txt
pyodbc==5.1.0
sqlalchemy==2.0.23
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv
jupyter==1.0.0
plotly==5.17.0
streamlit==1.28.1
requests==2.31.0
huggingface_hub==0.20.3
flask
pytest
pytest-mock
